jobs/pyproject.toml
[project]
name = "market-pulse-jobs"
version = "0.1.0"
description = "AlexStocks - Cron Jobs (ECS Fargate)"
requires-python = ">=3.11,<3.13"
dependencies = [
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dateutil>=2.8.0",
    "praw>=7.7.0",
    "prawcore>=2.4.0",
    "python-dotenv>=1.0.0",
    "vaderSentiment>=3.3.2",
    "transformers>=4.30.0,<4.40.0",
    "torch>=2.2.0,<2.3.0",
    "numpy>=1.24.0,<2.0.0",
    "tqdm>=4.66.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "yfinance>=0.2.66",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingest", "jobs"]
